Broke up NativeMethods into DLL-specific files
