+++ Cairo Desktop/Cairo Desktop/MenuBar.xaml.cs
@@ -373,11 +373,6 @@ namespace CairoDesktop
             return top;
         }
 
-        public NativeMethods.TrayHostSizeData GetMenuBarSizeData()
-        {
-            return new NativeMethods.TrayHostSizeData { edge = (int)appBarEdge, rc = new NativeMethods.Rect { Top = (int)(Top * dpiScale), Left = (int)(Left * dpiScale), Bottom = (int)((Top + Height) * dpiScale), Right = (int)((Left + Width) * dpiScale) } };
-        }
-
         protected override void CustomClosing()
         {
             if (WindowManager.Instance.IsSettingDisplays || Startup.IsShuttingDown)
