+++ Cairo Desktop/CairoDesktop.Interop/NativeMethods.Shell32.cs
@@ -1,0 +1,639 @@
+﻿using System;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+namespace CairoDesktop.Interop
+{
+    public partial class NativeMethods
+    {
+        const string Shell32_DllName = "shell32.dll";
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct APPBARDATA
+        {
+            public int cbSize;
+            public IntPtr hWnd;
+            public int uCallbackMessage;
+            public int uEdge;
+            public Rect rc;
+            public IntPtr lParam;
+        }
+
+        public enum ABMsg : int
+        {
+            ABM_NEW = 0,
+            ABM_REMOVE,
+            ABM_QUERYPOS,
+            ABM_SETPOS,
+            ABM_GETSTATE,
+            ABM_GETTASKBARPOS,
+            ABM_ACTIVATE,
+            ABM_GETAUTOHIDEBAR,
+            ABM_SETAUTOHIDEBAR,
+            ABM_WINDOWPOSCHANGED,
+            ABM_SETSTATE
+        }
+        public enum ABEdge : int
+        {
+            ABE_LEFT = 0,
+            ABE_TOP,
+            ABE_RIGHT,
+            ABE_BOTTOM
+        }
+
+        public enum AppBarNotifications
+        {
+            // Notifies an appbar that the taskbar's autohide or 
+            // always-on-top state has changed—that is, the user has selected 
+            // or cleared the "Always on top" or "Auto hide" check box on the
+            // taskbar's property sheet. 
+            StateChange = 0x00000000,
+            // Notifies an appbar when an event has occurred that may affect 
+            // the appbar's size and position. Events include changes in the
+            // taskbar's size, position, and visibility state, as well as the
+            // addition, removal, or resizing of another appbar on the same 
+            // side of the screen.
+            PosChanged = 0x00000001,
+            // Notifies an appbar when a full-screen application is opening or
+            // closing. This notification is sent in the form of an 
+            // application-defined message that is set by the ABM_NEW message. 
+            FullScreenApp = 0x00000002,
+            // Notifies an appbar that the user has selected the Cascade, 
+            // Tile Horizontally, or Tile Vertically command from the 
+            // taskbar's shortcut menu.
+            WindowArrange = 0x00000003
+        }
+
+        [DllImport(Shell32_DllName, CallingConvention = CallingConvention.StdCall)]
+        public static extern uint SHAppBarMessage(int dwMessage, ref APPBARDATA pData);
+
+        public const uint FILE_ATTRIBUTE_NORMAL = 0x80;
+        public const uint FILE_ATTRIBUTE_DIRECTORY = 0x10;
+
+
+        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
+        public struct SHFILEINFO
+        {
+            public IntPtr hIcon;
+            public int iIcon;
+            public uint dwAttributes;
+            [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 260)]
+            public string szDisplayName;
+            [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 80)]
+            public string szTypeName;
+        }
+
+        [Flags]
+        public enum SHGFI : int
+        {
+            /// <summary>get icon</summary>
+            Icon = 0x000000100,
+            /// <summary>get display name</summary>
+            DisplayName = 0x000000200,
+            /// <summary>get type name</summary>
+            TypeName = 0x000000400,
+            /// <summary>get attributes</summary>
+            Attributes = 0x000000800,
+            /// <summary>get icon location</summary>
+            IconLocation = 0x000001000,
+            /// <summary>return exe type</summary>
+            ExeType = 0x000002000,
+            /// <summary>get system icon index</summary>
+            SysIconIndex = 0x000004000,
+            /// <summary>put a link overlay on icon</summary>
+            LinkOverlay = 0x000008000,
+            /// <summary>show icon in selected state</summary>
+            Selected = 0x000010000,
+            /// <summary>get only specified attributes</summary>
+            Attr_Specified = 0x000020000,
+            /// <summary>get large icon</summary>
+            LargeIcon = 0x000000000,
+            /// <summary>get small icon</summary>
+            SmallIcon = 0x000000001,
+            /// <summary>get open icon</summary>
+            OpenIcon = 0x000000002,
+            /// <summary>get shell size icon</summary>
+            ShellIconSize = 0x000000004,
+            /// <summary>pszPath is a pidl</summary>
+            PIDL = 0x000000008,
+            /// <summary>use passed dwFileAttribute</summary>
+            UseFileAttributes = 0x000000010,
+            /// <summary>apply the appropriate overlays</summary>
+            AddOverlays = 0x000000020,
+            /// <summary>Get the index of the overlay in the upper 8 bits of the iIcon</summary>
+            OverlayIndex = 0x000000040,
+        }
+
+        [DllImport(Shell32_DllName, CharSet = CharSet.Unicode)]
+        public static extern IntPtr SHGetFileInfo(string pszPath, uint dwFileAttributes, ref SHFILEINFO psfi, uint cbSizeFileInfo, uint uFlags);
+
+        [DllImport(Shell32_DllName, CharSet = CharSet.Auto)]
+        public static extern bool ShellExecuteEx(ref SHELLEXECUTEINFO lpExecInfo);
+
+        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
+        public struct SHELLEXECUTEINFO
+        {
+            public int cbSize;
+            public uint fMask;
+            public IntPtr hwnd;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpVerb;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpFile;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpParameters;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpDirectory;
+            public int nShow;
+            public IntPtr hInstApp;
+            public IntPtr lpIDList;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpClass;
+            public IntPtr hkeyClass;
+            public uint dwHotKey;
+            public IntPtr hIcon;
+            public IntPtr hProcess;
+        }
+
+        public const uint SEE_MASK_INVOKEIDLIST = 12;
+
+        /// <summary>
+        /// Possible flags for the SHFileOperation method.
+        /// </summary>
+        [Flags]
+        public enum FileOperationFlags : ushort
+        {
+            /// <summary>
+            /// Do not show a dialog during the process
+            /// </summary>
+            FOF_SILENT = 0x0004,
+            /// <summary>
+            /// Do not ask the user to confirm selection
+            /// </summary>
+            FOF_NOCONFIRMATION = 0x0010,
+            /// <summary>
+            /// Delete the file to the recycle bin.  (Required flag to send a file to the bin
+            /// </summary>
+            FOF_ALLOWUNDO = 0x0040,
+            /// <summary>
+            /// Do not show the names of the files or folders that are being recycled.
+            /// </summary>
+            FOF_SIMPLEPROGRESS = 0x0100,
+            /// <summary>
+            /// Surpress errors, if any occur during the process.
+            /// </summary>
+            FOF_NOERRORUI = 0x0400,
+            /// <summary>
+            /// Warn if files are too big to fit in the recycle bin and will need
+            /// to be deleted completely.
+            /// </summary>
+            FOF_WANTNUKEWARNING = 0x4000,
+        }
+
+        /// <summary>
+        /// File Operation Function Type for SHFileOperation
+        /// </summary>
+        public enum FileOperationType : uint
+        {
+            /// <summary>
+            /// Move the objects
+            /// </summary>
+            FO_MOVE = 0x0001,
+            /// <summary>
+            /// Copy the objects
+            /// </summary>
+            FO_COPY = 0x0002,
+            /// <summary>
+            /// Delete (or recycle) the objects
+            /// </summary>
+            FO_DELETE = 0x0003,
+            /// <summary>
+            /// Rename the object(s)
+            /// </summary>
+            FO_RENAME = 0x0004,
+        }
+
+        /// <summary>
+        /// SHFILEOPSTRUCT for SHFileOperation from COM
+        /// </summary>
+        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
+        public struct SHFILEOPSTRUCT
+        {
+
+            public IntPtr hwnd;
+            [MarshalAs(UnmanagedType.U4)]
+            public FileOperationType wFunc;
+            public string pFrom;
+            public string pTo;
+            public FileOperationFlags fFlags;
+            [MarshalAs(UnmanagedType.Bool)]
+            public bool fAnyOperationsAborted;
+            public IntPtr hNameMappings;
+            public string lpszProgressTitle;
+        }
+
+        [DllImport(Shell32_DllName, CharSet = CharSet.Auto)]
+        public static extern int SHFileOperation(ref SHFILEOPSTRUCT FileOp);
+
+        public const uint NIN_SELECT = 0x400;
+        public const uint NIN_POPUPOPEN = 0x406;
+        public const uint NIN_POPUPCLOSE = 0x407;
+
+        /// <summary>
+        /// Numerical values of the NIM_* messages represented as an enumeration.
+        /// </summary>
+        public enum NIM : uint
+        {
+            /// <summary>
+            /// Add a new icon.
+            /// </summary>
+            NIM_ADD = 0,
+
+            /// <summary>
+            /// Modify an existing icon.
+            /// </summary>
+            NIM_MODIFY = 1,
+
+            /// <summary>
+            /// Delete an icon.
+            /// </summary>
+            NIM_DELETE = 2,
+
+            /// <summary>
+            /// Shell v5 and above - Return focus to the notification area.
+            /// </summary>
+            NIM_SETFOCUS = 3,
+
+            /// <summary>
+            /// Shell v4 and above - Instructs the taskbar to behave accordingly based on the version (uVersion) set in the notifiyicondata struct.
+            /// </summary>
+            NIM_SETVERSION = 4
+        }
+
+        /// <summary>
+        /// Shell_NotifyIcon flags.  NIF_*
+        /// </summary>
+        [Flags]
+        public enum NIF : uint
+        {
+            MESSAGE = 0x0001,
+            ICON = 0x0002,
+            TIP = 0x0004,
+            STATE = 0x0008,
+            INFO = 0x0010,
+            GUID = 0x0020,
+
+            /// <summary>
+            /// Vista only.
+            /// </summary>
+            REALTIME = 0x0040,
+            /// <summary>
+            /// Vista only.
+            /// </summary>
+            SHOWTIP = 0x0080,
+
+            XP_MASK = STATE | INFO | GUID,
+            VISTA_MASK = REALTIME | SHOWTIP,
+        }
+
+        /// <summary>
+        /// Notify icon data structure type
+        /// </summary>
+        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
+        public struct NOTIFYICONDATA
+        {
+            public int cbSize;
+            public uint hWnd;
+            public uint uID;
+            public NIF uFlags;
+            public uint uCallbackMessage;
+            public uint hIcon;
+            [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 128)]
+            public string szTip;
+            public int dwState;
+            public int dwStateMask;
+            [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 256)]
+            public string szInfo;
+            public uint uVersion;  // used with NIM_SETVERSION, values 0, 3 and 4
+            [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 64)]
+            public string szInfoTitle;
+            public uint dwInfoFlags;
+            public Guid guidItem;
+            public uint hBalloonIcon;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct SHELLTRAYDATA
+        {
+            public int dwUnknown;
+            public uint dwMessage;
+            public NOTIFYICONDATA nid;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct WINNOTIFYICONIDENTIFIER
+        {
+            public int dwMagic;
+            public int dwMessage;
+            public int cbSize;
+            public int dwPadding;
+            public uint hWnd;
+            public uint uID;
+            public Guid guidItem;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct APPBARDATAV2
+        {
+            public int cbSize;
+            public uint hWnd;
+            public uint uCallbackMessage;
+            public uint uEdge;
+            public Rect rc;
+            public int lParam;
+            public int dw64BitAlign;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct APPBARMSGDATAV3
+        {
+            public APPBARDATAV2 abd;
+            public int dwMessage;
+            public int dwPadding1;
+            public uint hSharedMemory;
+            public int dwPadding2;
+            public int dwSourceProcessId;
+            public int dwPadding3;
+        }
+
+        [StructLayout(LayoutKind.Sequential, Pack = 4)]
+        public struct PROPERTYKEY
+        {
+            public Guid fmtid;
+            public uint pid;
+        }
+
+        [ComImport, Guid("886D8EEB-8CF2-4446-8D02-CDBA1DBDCF99"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+        public interface IPropertyStore
+        {
+            [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+            void GetCount([Out] out uint cProps);
+
+            [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+            void GetAt([In] uint iProp, out PROPERTYKEY pkey);
+
+            [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+            void GetValue([In] ref PROPERTYKEY key, out PropVariant pv);
+
+            [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+            void SetValue([In] ref PROPERTYKEY key, [In] ref PropVariant pv);
+
+            [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime)]
+            void Commit();
+        }
+
+        [DllImport(Shell32_DllName, SetLastError = true)]
+        public static extern int SHGetPropertyStoreForWindow(IntPtr handle, ref Guid riid, out IPropertyStore propertyStore);
+
+        [Flags()]
+        public enum RunFileDialogFlags : uint
+        {
+
+            /// <summary>
+            /// Don't use any of the flags (only works alone)
+            /// </summary>
+            None = 0x0000,
+
+            /// <summary>
+            /// Removes the browse button
+            /// </summary>
+            NoBrowse = 0x0001,
+
+            /// <summary>
+            /// No default item selected
+            /// </summary>
+            NoDefault = 0x0002,
+
+            /// <summary>
+            /// Calculates the working directory from the file name
+            /// </summary>
+            CalcDirectory = 0x0004,
+
+            /// <summary>
+            /// Removes the edit box label
+            /// </summary>
+            NoLabel = 0x0008,
+
+            /// <summary>
+            /// Removes the separate memory space checkbox (Windows NT only)
+            /// </summary>
+            NoSeperateMemory = 0x0020
+        }
+
+        [DllImport(Shell32_DllName, CharSet = CharSet.Auto, EntryPoint = "#61", SetLastError = true)]
+        public static extern bool SHRunFileDialog(IntPtr hwndOwner,
+            IntPtr hIcon,
+            string lpszPath,
+            string lpszDialogTitle,
+            string lpszDialogTextBody,
+            RunFileDialogFlags uflags);
+
+        public struct IMAGELISTDRAWPARAMS
+        {
+            public int cbSize;
+            public IntPtr himl;
+            public int i;
+            public IntPtr hdcDst;
+            public int x;
+            public int y;
+            public int cx;
+            public int cy;
+            public int xBitmap;        // x offest from the upperleft of bitmap
+            public int yBitmap;        // y offset from the upperleft of bitmap
+            public int rgbBk;
+            public int rgbFg;
+            public int fStyle;
+            public int dwRop;
+            public int fState;
+            public int Frame;
+            public int crEffect;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct IMAGEINFO
+        {
+            public IntPtr hbmImage;
+            public IntPtr hbmMask;
+            public int Unused1;
+            public int Unused2;
+            public Rect rcImage;
+        }
+
+        [ComImportAttribute()]
+        [GuidAttribute("46EB5926-582E-4017-9FDF-E8998DAA0950")]
+        [InterfaceTypeAttribute(ComInterfaceType.InterfaceIsIUnknown)]
+        public interface IImageList
+        {
+            [PreserveSig]
+            int Add(
+                IntPtr hbmImage,
+                IntPtr hbmMask,
+                ref int pi);
+
+            [PreserveSig]
+            int ReplaceIcon(
+                int i,
+                IntPtr hicon,
+                ref int pi);
+
+            [PreserveSig]
+            int SetOverlayImage(
+                int iImage,
+                int iOverlay);
+
+            [PreserveSig]
+            int Replace(
+                int i,
+                IntPtr hbmImage,
+                IntPtr hbmMask);
+
+            [PreserveSig]
+            int AddMasked(
+                IntPtr hbmImage,
+                int crMask,
+                ref int pi);
+
+            [PreserveSig]
+            int Draw(
+                ref IMAGELISTDRAWPARAMS pimldp);
+
+            [PreserveSig]
+            int Remove(
+            int i);
+
+            [PreserveSig]
+            int GetIcon(
+                int i,
+                int flags,
+                ref IntPtr picon);
+
+            [PreserveSig]
+            int GetImageInfo(
+                int i,
+                ref IMAGEINFO pImageInfo);
+
+            [PreserveSig]
+            int Copy(
+                int iDst,
+                IImageList punkSrc,
+                int iSrc,
+                int uFlags);
+
+            [PreserveSig]
+            int Merge(
+                int i1,
+                IImageList punk2,
+                int i2,
+                int dx,
+                int dy,
+                ref Guid riid,
+                ref IntPtr ppv);
+
+            [PreserveSig]
+            int Clone(
+                ref Guid riid,
+                ref IntPtr ppv);
+
+            [PreserveSig]
+            int GetImageRect(
+                int i,
+                ref Rect prc);
+
+            [PreserveSig]
+            int GetIconSize(
+                ref int cx,
+                ref int cy);
+
+            [PreserveSig]
+            int SetIconSize(
+                int cx,
+                int cy);
+
+            [PreserveSig]
+            int GetImageCount(
+            ref int pi);
+
+            [PreserveSig]
+            int SetImageCount(
+                int uNewCount);
+
+            [PreserveSig]
+            int SetBkColor(
+                int clrBk,
+                ref int pclr);
+
+            [PreserveSig]
+            int GetBkColor(
+                ref int pclr);
+
+            [PreserveSig]
+            int BeginDrag(
+                int iTrack,
+                int dxHotspot,
+                int dyHotspot);
+
+            [PreserveSig]
+            int EndDrag();
+
+            [PreserveSig]
+            int DragEnter(
+                IntPtr hwndLock,
+                int x,
+                int y);
+
+            [PreserveSig]
+            int DragLeave(
+                IntPtr hwndLock);
+
+            [PreserveSig]
+            int DragMove(
+                int x,
+                int y);
+
+            [PreserveSig]
+            int SetDragCursorImage(
+                ref IImageList punk,
+                int iDrag,
+                int dxHotspot,
+                int dyHotspot);
+
+            [PreserveSig]
+            int DragShowNolock(
+                int fShow);
+
+            [PreserveSig]
+            int GetDragImage(
+                ref POINT ppt,
+                ref POINT pptHotspot,
+                ref Guid riid,
+                ref IntPtr ppv);
+
+            [PreserveSig]
+            int GetItemFlags(
+                int i,
+                ref int dwFlags);
+
+            [PreserveSig]
+            int GetOverlayImage(
+                int iOverlay,
+                ref int piIndex);
+        };
+
+        [DllImport(Shell32_DllName, EntryPoint = "#727")]
+        public extern static int SHGetImageList(
+            int iImageList,
+            ref Guid riid,
+            out IImageList ppv
+            );
+    }
+}
